refactor(monolith): split build_topology into observability/topology_query
Move build_topology and its ClickHouse helpers (_ch_scalar, _ch_rows,
_query_node, _query_edge, the semaphore and retry constants) out of
home/observability/router.py into a new home/observability/topology_query.py.

The two public read endpoints (/stats and /topology) only read precomputed
Postgres snapshot rows, so router.py no longer imports the ClickHouse client,
SLO math, or topology config. rollup.py (the private writer) now imports
build_topology from topology_query. This shrinks the public import closure for
app.main_public (ADR 004 Layer 1+4): no behavior change for the private app.
